Reorganizes cli tests to cli folder
