+++ tests/vhs/test_cli_login/test_login.yaml
@@ -1,6 +1,6 @@
 interactions:
 - request:
-    body: null
+    body: grant_type=password&client_id=FAKE_CLIENT_ID&client_secret=FAKE_CLIENT_SECRET
     headers:
       Accept:
       - '*/*'
@@ -8,28 +8,31 @@ interactions:
       - gzip, deflate
       Connection:
       - keep-alive
+      Content-Length:
+      - '109'
+      Content-Type:
+      - application/x-www-form-urlencoded
       User-Agent:
       - python-requests/2.30.0
       authorization:
       - Bearer DUMMY
-    method: GET
-    uri: https://api.cimediacloud.com/workspaces/f44132c2393d470c88b9884f45877ebb/search?query=i+am+not+a+guid&kind=asset
+    method: POST
+    uri: https://api.cimediacloud.com/oauth2/token
   response:
     body:
-      string: '{"limit": 50, "offset": 0, "count": 0, "kind": "Asset", "type": "All",
-        "query": "i am not a guid", "items": [], "order": {"by": "Relevance", "direction":
-        "Desc"}}'
+      string: '{"access_token": "DUMMY_ACCESS_TOKEN", "expires_in": 86400, "token_type":
+        "bearer", "refresh_token": "DUMMY_REFRESH_TOKEN"}'
     headers:
       Cache-Control:
       - no-cache
       Connection:
       - keep-alive
       Content-Length:
-      - '144'
+      - '143'
       Content-Type:
       - application/json; charset=utf-8
       Date:
-      - Thu, 11 May 2023 16:34:38 GMT
+      - Thu, 11 May 2023 17:28:02 GMT
       Expires:
       - '-1'
       Pragma:
@@ -38,12 +41,6 @@ interactions:
       - max-age=86400; includeSubDomains
       X-Frame-Options:
       - deny
-      X-RateLimit-Limit:
-      - '5000'
-      X-RateLimit-Remaining:
-      - '4984'
-      X-RateLimit-Reset:
-      - '1683822900'
     status:
       code: 200
       message: OK
