+++ tests/vhs/test_cli_search/test_empty_search.yaml
@@ -1,6 +1,6 @@
 interactions:
 - request:
-    body: grant_type=password&client_id=FAKE_CLIENT_ID&client_secret=FAKE_CLIENT_SECRET
+    body: null
     headers:
       Accept:
       - '*/*'
@@ -8,31 +8,28 @@ interactions:
       - gzip, deflate
       Connection:
       - keep-alive
-      Content-Length:
-      - '53'
-      Content-Type:
-      - application/x-www-form-urlencoded
       User-Agent:
       - python-requests/2.30.0
       authorization:
       - Bearer DUMMY
-    method: POST
-    uri: https://api.cimediacloud.com/oauth2/token
+    method: GET
+    uri: https://api.cimediacloud.com/workspaces/f44132c2393d470c88b9884f45877ebb/search?query=i+am+not+a+guid&kind=asset
   response:
     body:
-      string: '{"error": "invalid_client", "error_description": "Invalid client id
-        and client secret combination."}'
+      string: '{"limit": 50, "offset": 0, "count": 0, "kind": "Asset", "type": "All",
+        "query": "i am not a guid", "items": [], "order": {"by": "Relevance", "direction":
+        "Desc"}}'
     headers:
       Cache-Control:
       - no-cache
       Connection:
       - keep-alive
       Content-Length:
-      - '97'
+      - '144'
       Content-Type:
       - application/json; charset=utf-8
       Date:
-      - Thu, 11 May 2023 16:34:38 GMT
+      - Thu, 11 May 2023 17:26:36 GMT
       Expires:
       - '-1'
       Pragma:
@@ -41,7 +38,13 @@ interactions:
       - max-age=86400; includeSubDomains
       X-Frame-Options:
       - deny
+      X-RateLimit-Limit:
+      - '5000'
+      X-RateLimit-Remaining:
+      - '4969'
+      X-RateLimit-Reset:
+      - '1683826020'
     status:
-      code: 401
-      message: Unauthorized
+      code: 200
+      message: OK
 version: 1
