--- tests/vhs/tests.test_cli/test_guid_search.yaml
@@ -1,76 +1,0 @@
-interactions:
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - python-requests/2.30.0
-      authorization:
-      - Bearer DUMMY
-    method: GET
-    uri: https://api.cimediacloud.com/workspaces/f44132c2393d470c88b9884f45877ebb/search?query=pb-aacip-e4308199588&kind=asset
-  response:
-    body:
-      string: '{"limit": 50, "offset": 0, "count": 1, "kind": "Asset", "type": "All",
-        "query": "pb-aacip-e4308199588", "items": [{"id": "554544ceaf6b4c94a4a06cee5bc1f39f",
-        "name": "cpb-aacip-e4308199588.txt", "size": 16, "createdOn": "2021-08-17T15:13:59.841Z",
-        "thumbnails": [{"type": "medium", "location": "https://cdn01.cimediacloud.com/cifiles/554544ceaf6b4c94a4a06cee5bc1f39f/renditions/thumbnail-medium.jpg?Policy=eyJTdGF0ZW1lbnQiOlt7IlJlc291cmNlIjoiKjovL2NkbjAxLmNpbWVkaWFjbG91ZC5jb20vY2lmaWxlcy81NTQ1NDRjZWFmNmI0Yzk0YTRhMDZjZWU1YmMxZjM5Zi9yZW5kaXRpb25zL3RodW1ibmFpbC1tZWRpdW0uanBnIiwiQ29uZGl0aW9uIjp7IklwQWRkcmVzcyI6eyJBV1M6U291cmNlSXAiOiIwLjAuMC4wLzAifSwiRGF0ZUxlc3NUaGFuIjp7IkFXUzpFcG9jaFRpbWUiOjE3NDY4OTQ4Nzl9fX1dfQ__&Signature=BQBkYd1dPWJEWz2IPRPlUzm5iyFu1b528KJEnq92Z7TdDdOR-1UUOFrnXLecA6ejWBzFULKevwklosaIjRzlFX7jApLBPYo7PDbAVX4-YlTksYD~0VZCMo5kenauQIH62GgxoJcnXP6dUcDztAZu-xbpZtajaCOcUG6i12IzDdM_&Key-Pair-Id=APKAJNUSFH4IKQRZ44WA",
-        "size": 3554, "width": 433, "height": 560, "isExternal": false}, {"type":
-        "large", "location": "https://cdn01.cimediacloud.com/cifiles/554544ceaf6b4c94a4a06cee5bc1f39f/renditions/thumbnail-large.jpg?Policy=eyJTdGF0ZW1lbnQiOlt7IlJlc291cmNlIjoiKjovL2NkbjAxLmNpbWVkaWFjbG91ZC5jb20vY2lmaWxlcy81NTQ1NDRjZWFmNmI0Yzk0YTRhMDZjZWU1YmMxZjM5Zi9yZW5kaXRpb25zL3RodW1ibmFpbC1sYXJnZS5qcGciLCJDb25kaXRpb24iOnsiSXBBZGRyZXNzIjp7IkFXUzpTb3VyY2VJcCI6IjAuMC4wLjAvMCJ9LCJEYXRlTGVzc1RoYW4iOnsiQVdTOkVwb2NoVGltZSI6MTc0Njg5NDg3OX19fV19&Signature=IiG9OuICHn6A5Wm5pAAkNw6S~fymfR7TIc63ht82duhRefVARBSwYrmMxfOvt-gCob3eg2glPVgAjYoLsrXtgAAz6nOk-5xK8bM6KvzUPgdv7JO9dpTy9ukXpt7NJ0arDMMac7PGV~UlzfB0Pm9batZ5UORXSIYB3KC2gRc7wHo_&Key-Pair-Id=APKAJNUSFH4IKQRZ44WA",
-        "size": 7208, "width": 612, "height": 792, "isExternal": false}, {"type":
-        "2000px", "location": "https://cdn01.cimediacloud.com/cifiles/554544ceaf6b4c94a4a06cee5bc1f39f/renditions/thumbnail-2000px.jpg?Policy=eyJTdGF0ZW1lbnQiOlt7IlJlc291cmNlIjoiKjovL2NkbjAxLmNpbWVkaWFjbG91ZC5jb20vY2lmaWxlcy81NTQ1NDRjZWFmNmI0Yzk0YTRhMDZjZWU1YmMxZjM5Zi9yZW5kaXRpb25zL3RodW1ibmFpbC0yMDAwcHguanBnIiwiQ29uZGl0aW9uIjp7IklwQWRkcmVzcyI6eyJBV1M6U291cmNlSXAiOiIwLjAuMC4wLzAifSwiRGF0ZUxlc3NUaGFuIjp7IkFXUzpFcG9jaFRpbWUiOjE3NDY4OTQ4Nzl9fX1dfQ__&Signature=I1b1A-rOEXWBu2gRSqURsBK8CrnFQPQsIkAX-d8ZBj8zjIBCHiqZMC2eGKkjnAqAiggepFZ5WNfvDBlgmi2Qd~8ri9M20n8t4kBxHekliN9VsfS27vwBATUVze2nNGVg7BybYoawjHJwc~9UfP~xbQ0RhRe2KV116SCWKpF51pg_&Key-Pair-Id=APKAJNUSFH4IKQRZ44WA",
-        "size": 7208, "width": 612, "height": 792, "isExternal": false}, {"type":
-        "small", "location": "https://cdn01.cimediacloud.com/cifiles/554544ceaf6b4c94a4a06cee5bc1f39f/renditions/thumbnail.jpg?Policy=eyJTdGF0ZW1lbnQiOlt7IlJlc291cmNlIjoiKjovL2NkbjAxLmNpbWVkaWFjbG91ZC5jb20vY2lmaWxlcy81NTQ1NDRjZWFmNmI0Yzk0YTRhMDZjZWU1YmMxZjM5Zi9yZW5kaXRpb25zL3RodW1ibmFpbC5qcGciLCJDb25kaXRpb24iOnsiSXBBZGRyZXNzIjp7IkFXUzpTb3VyY2VJcCI6IjAuMC4wLjAvMCJ9LCJEYXRlTGVzc1RoYW4iOnsiQVdTOkVwb2NoVGltZSI6MTc0Njg5NDg3OX19fV19&Signature=TCLeUdCAGmeSlO0VSeVbrVPf6KDmA2u3ILM0DNunyQiCeOpzGn8PIP2053s9SQugquxsN3CdhCIU8QCOHg6bnorsvI1RonPm6cQFZ8mXhcvrpPLZv62qn~UI2kl72CnGPr5bIT2IiwHGpIh7A5Ny37JQA-Mgc4RprEu4ByuJ45M_&Key-Pair-Id=APKAJNUSFH4IKQRZ44WA",
-        "size": 1038, "width": 193, "height": 250, "isExternal": false}, {"type":
-        "standard", "location": "https://cdn01.cimediacloud.com/cifiles/554544ceaf6b4c94a4a06cee5bc1f39f/renditions/thumbnail.jpg?Policy=eyJTdGF0ZW1lbnQiOlt7IlJlc291cmNlIjoiKjovL2NkbjAxLmNpbWVkaWFjbG91ZC5jb20vY2lmaWxlcy81NTQ1NDRjZWFmNmI0Yzk0YTRhMDZjZWU1YmMxZjM5Zi9yZW5kaXRpb25zL3RodW1ibmFpbC5qcGciLCJDb25kaXRpb24iOnsiSXBBZGRyZXNzIjp7IkFXUzpTb3VyY2VJcCI6IjAuMC4wLjAvMCJ9LCJEYXRlTGVzc1RoYW4iOnsiQVdTOkVwb2NoVGltZSI6MTc0Njg5NDg3OX19fV19&Signature=TCLeUdCAGmeSlO0VSeVbrVPf6KDmA2u3ILM0DNunyQiCeOpzGn8PIP2053s9SQugquxsN3CdhCIU8QCOHg6bnorsvI1RonPm6cQFZ8mXhcvrpPLZv62qn~UI2kl72CnGPr5bIT2IiwHGpIh7A5Ny37JQA-Mgc4RprEu4ByuJ45M_&Key-Pair-Id=APKAJNUSFH4IKQRZ44WA",
-        "size": 1038, "width": 193, "height": 250}], "proxies": [{"type": "document-pdf",
-        "location": "https://cdn01.cimediacloud.com/cifiles/554544ceaf6b4c94a4a06cee5bc1f39f/renditions/document-pdf.pdf?u=16a555d05c60439a8243a3d6c349fd8b&a=554544ceaf6b4c94a4a06cee5bc1f39f&ct=abb0668d5b054200b6a89c3826024e42&et=stream&Policy=eyJTdGF0ZW1lbnQiOlt7IlJlc291cmNlIjoiKjovL2NkbjAxLmNpbWVkaWFjbG91ZC5jb20vY2lmaWxlcy81NTQ1NDRjZWFmNmI0Yzk0YTRhMDZjZWU1YmMxZjM5Zi9yZW5kaXRpb25zL2RvY3VtZW50LXBkZi5wZGY~dT0xNmE1NTVkMDVjNjA0MzlhODI0M2EzZDZjMzQ5ZmQ4YiZhPTU1NDU0NGNlYWY2YjRjOTRhNGEwNmNlZTViYzFmMzlmJmN0PWFiYjA2NjhkNWIwNTQyMDBiNmE4OWMzODI2MDI0ZTQyJmV0PXN0cmVhbSIsIkNvbmRpdGlvbiI6eyJJcEFkZHJlc3MiOnsiQVdTOlNvdXJjZUlwIjoiMC4wLjAuMC8wIn0sIkRhdGVMZXNzVGhhbiI6eyJBV1M6RXBvY2hUaW1lIjoxNjgzODY2MDc5fX19XX0_&Signature=fcIxIbUZGup8eOFjQVWWSdKMtCFlJF8gb3vLIX4fklkQ6WkS9PQpY3Lqv2prMqrcL347wrGDFkznL40j2s-zsvDn0-5-owHhczFJkVW36aKuoib~gTl1JYfn4ndNuFbWUNBwWlwNDwE8ZR18QvMYo5XXHiH59obSmq3o4L3vbaI_&Key-Pair-Id=APKAJNUSFH4IKQRZ44WA",
-        "size": 7172, "isExternal": false}], "type": "Document", "format": "txt",
-        "status": "Complete", "lastActivityOn": "2023-05-10T16:44:14.834Z", "modifiedOn":
-        "2021-08-17T15:14:43.43Z", "md5Checksum": "23211babeacb8adf336194eaca77f510",
-        "clipSource": {}, "technicalMetadata": {}, "archiveStatus": "Not archived",
-        "restoreStatus": "Not restored", "uploadCompleteDate": "2021-08-17T15:13:59.966Z",
-        "isTrashed": false, "isFavorite": false, "isExternal": false, "uploadTransferType":
-        "SinglepartHttp", "folder": {"id": "d870616520634dbc8eb8caec141188fa", "name":
-        "Workspace"}, "workspace": {"id": "f44132c2393d470c88b9884f45877ebb", "name":
-        "My Workspace", "class": "Personal"}, "network": {"id": "70c6959f648a428096c88906b6655773",
-        "name": "AAPB''s Network", "class": "Personal"}, "filmstrips": [], "waveforms":
-        [], "mediaAnalyses": [], "acquisitionSource": {"name": "American Archive"},
-        "totalFolderCount": 1, "isTimedTextMatch": false, "kind": "Asset", "createdBy":
-        {"id": "16a555d05c60439a8243a3d6c349fd8b", "name": "AAPB Notifications", "email":
-        "aapb_notifications@wgbh.org"}}], "order": {"by": "Relevance", "direction":
-        "Desc"}}'
-    headers:
-      Cache-Control:
-      - no-cache
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '5785'
-      Content-Type:
-      - application/json; charset=utf-8
-      Date:
-      - Thu, 11 May 2023 16:34:39 GMT
-      Expires:
-      - '-1'
-      Pragma:
-      - no-cache
-      Strict-Transport-Security:
-      - max-age=86400; includeSubDomains
-      X-Frame-Options:
-      - deny
-      X-RateLimit-Limit:
-      - '5000'
-      X-RateLimit-Remaining:
-      - '4983'
-      X-RateLimit-Reset:
-      - '1683822900'
-    status:
-      code: 200
-      message: OK
-version: 1
